feat(page): added page layout to handle basic page layouts
