+++ src/components/BackButton/index.js
@@ -1,0 +1,20 @@
+/** @jsx jsx */
+import { jsx } from '@emotion/core'
+
+import { Space } from 'lib/theme'
+
+import { IconButton } from 'components/Button'
+
+import { ReactComponent as Back } from 'static/icons/back-circle.svg'
+
+import styles from './BackButton.styles'
+
+export const BackButton = ({ onClick }) => (
+  <div css={styles.root}>
+    <IconButton onClick={onClick}>
+      <Back />
+    </IconButton>
+  </div>
+)
+
+export default BackButton
