+++ src/components/layouts/Page/Page.styles.js
@@ -1,13 +1,18 @@
 import { css } from '@emotion/core'
-import { Color, Space } from 'lib/theme'
+
+import { Space } from 'lib/theme'
 
 const styles = {
   root: css({
-    color: Color.GRAY,
     display: 'flex',
     flex: 1,
     flexDirection: 'column',
-    padding: `${Space.S30}px ${Space.S40}px`,
+  }),
+  content: css({
+    display: 'flex',
+    flex: 1,
+    flexDirection: 'column',
+    padding: `0 ${Space.S40}px`,
   }),
 }
