+++ src/index.css
@@ -3,6 +3,7 @@
 html {
   height: 100%;
 }
+
 body {
   height: 100%;
   font-family: 'Inter', Helvetica, Arial, sans-serif;
@@ -15,5 +16,7 @@ body {
 }
 
 .App {
-  height: 100%;
+  display: flex;
+  flex-direction: column;
+  min-height: 100%;
 }
