+++ src/pages/request.js
@@ -1,22 +1,16 @@
 /** @jsx jsx */
-import React, { Fragment } from 'react';
+import React from 'react';
 import { jsx } from '@emotion/core';
 import { withRouter } from 'react-router-dom';
-import Box from 'components/Box';
+import Page from 'components/layouts/Page';
 import FindFacility from 'containers/FindFacility';
-import BackButton from 'components/BackButton';
-import Header from 'components/Header';
 
 class Request extends React.Component {
   render() {
     return (
-      <Fragment>
-        <Header />
-        <Box>
-          <BackButton />
-          <FindFacility {...this.props} />
-        </Box>
-      </Fragment>
+      <Page>
+        <FindFacility {...this.props} />
+      </Page>
     );
   }
 }
